Refactor quantization files
